Download data at the municipality level

`municipios()` downloads, formats and imports to the environment the electoral results data of the selected election at the municipality level.

municipios(tipo_eleccion, anno, mes, distritos = FALSE)

Arguments

tipo_eleccion

The type of choice you want to download. The accepted values are "congreso", "senado", "europeas" o "municipales".

anno

The year of the election in YYYY format.

mes

The month of the election in MM format.

distritos

Should district level results be returned when available? The default is FALSE. Please be aware when summarizing the data that districts = TRUE will return separate rows for the total municipal level and for each of the districts.

Value

Dataframe with the electoral results data at the municipality level.

Examples

data <- municipios(tipo_eleccion = "congreso", anno = "2019", mes = "11")
#> Downloading https://infoelectoral.interior.gob.es/estaticos/docxl/apliextr/02201911_MUNI.zip
str(data)
#> 'data.frame':	99118 obs. of  34 variables:
#>  $ tipo_eleccion            : chr  "02" "02" "02" "02" ...
#>  $ anno                     : chr  "2019" "2019" "2019" "2019" ...
#>  $ mes                      : chr  "11" "11" "11" "11" ...
#>  $ vuelta                   : chr  "1" "1" "1" "1" ...
#>  $ codigo_ccaa              : chr  "01" "01" "01" "01" ...
#>  $ codigo_provincia         : chr  "04" "04" "04" "04" ...
#>  $ codigo_municipio         : chr  "001" "001" "001" "001" ...
#>  $ municipio                : chr  "Abla" "Abla" "Abla" "Abla" ...
#>  $ codigo_distrito          : chr  "99" "99" "99" "99" ...
#>  $ codigo_distrito_electoral: chr  "0" "0" "0" "0" ...
#>  $ codigo_partido_judicial  : chr  "013" "013" "013" "013" ...
#>  $ codigo_diputacion        : chr  "013" "013" "013" "013" ...
#>  $ codigo_comarca           : chr  "000" "000" "000" "000" ...
#>  $ poblacion_derecho        : num  1235 1235 1235 1235 1235 ...
#>  $ numero_mesas             : num  2 2 2 2 2 2 2 2 2 2 ...
#>  $ censo_ine                : num  1002 1002 1002 1002 1002 ...
#>  $ censo_escrutinio         : num  1002 1002 1002 1002 1002 ...
#>  $ censo_cere               : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ votantes_cere            : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ participacion_1          : num  393 393 393 393 393 393 393 393 393 393 ...
#>  $ participacion_2          : num  588 588 588 588 588 588 588 588 588 588 ...
#>  $ votos_blancos            : num  3 3 3 3 3 3 3 3 3 3 ...
#>  $ votos_nulos              : num  7 7 7 7 7 7 7 7 7 7 ...
#>  $ votos_candidaturas       : num  707 707 707 707 707 707 707 707 707 707 ...
#>  $ numero_concejales        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ codigo_partido_nacional  : chr  "000009" "000018" "000023" "000055" ...
#>  $ codigo_partido_autonomia : chr  "000009" "000018" "000023" "000055" ...
#>  $ codigo_partido_provincia : chr  "000009" "000018" "000023" "000055" ...
#>  $ codigo_partido           : chr  "000009" "000018" "000023" "000055" ...
#>  $ denominacion             : chr  "ANDALUCÍA POR SÍ" "CIUDADANOS-PARTIDO DE LA CIUDADANÍA" "ESCAÑOS EN BLANCO" "PARTIDO ANIMALISTA CONTRA EL MALTRATO ANIMAL" ...
#>  $ siglas                   : chr  "AxSÍ" "Cs" "EB" "PACMA" ...
#>  $ votos                    : num  0 47 2 3 0 30 193 310 0 0 ...
#>  $ datos_oficiales          : chr  "S" "S" "S" "S" ...
#>  $ concejales_obtenidos     : num  0 0 0 0 0 0 0 0 0 0 ...