candidatos

Esta función descarga los datos de los candidatos de las listas electorales de las elecciones seleccionadas, los formatea, y los importa al espacio de trabajo.

Uso

candidatos(tipo_eleccion, anno, mes)

Argumentos

tipo_eleccion

El tipo de eleccion que se quiere descargar. Los valores aceptados por ahora son "municipales" o "generales".

anno

El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").

mes

El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).

Valor

Dataframe con los datos de candidatos.